k32w0: pdm: overwrite ramStorageEnsureBlockConsistency
A key shall not span across two PDM id regions to avoid
having corrupted data in some corner cases.

A dummy key is used to fill in the space that could not fit
a certain key at the end of a PDM region. The key will be
written in the next PDM region, ensuring it does not span
across two PDM regions.

Signed-off-by: Marius Tache <marius.tache@nxp.com>