rt1060: enable mac filtering by default
For RT1170 we've already enable the MAC filtering feature for OT,
which helps running tests in "open air" as opposed to a Faraday cage.
In order to expedite the conformance testing, it's easier to run in
"open air", but MAC filtering needs to be enabled so other devices do
not mess up the tests.

As such, set OPENTHREAD_CONFIG_MAC_FILTER_ENABLE to 1 also for RT1060.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>